Early error for arguments in initializer (#56)
At the November 2016 TC39 meeting as well as previous meetings,
the committee discussed potential confusion from the value of
`arguments` inside a class field initializer. This PR creates an
early error for such a reference, including inside eval. I had
previously thought that an early error for a case like this would
be unusual and a runtime error would therefore be preferable, but
@adamk pointed out that such an error is analogous to errors for
new.target in the wrong location; the specification here is
analogous as well.